refactor: do the same in python
