+++ source/modulo_components/test/python/test_component_interface_parameters.py
@@ -3,7 +3,7 @@ import rclpy.exceptions
 import rclpy.node
 import state_representation as sr
 from modulo_components.component_interface import ComponentInterface
-from modulo_utils.exceptions import ParameterError
+from modulo_core.exceptions import ParameterError
 from rcl_interfaces.msg import SetParametersResult
