--- source/modulo_core/modulo_core/exceptions/__init__.py
@@ -1,1 +1,0 @@
-from modulo_core.exceptions.core_exceptions import CoreError, MessageTranslationError, ParameterTranslationError
